requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.2.15
    # via
    #   -r requirements.in
    #   django-basicauth
    #   django-braces
    #   django-extensions
    #   djangorestframework
    #   drf-yasg
django-basicauth==0.5.3
    # via -r requirements.in
django-braces==1.15.0
    # via -r requirements.in
django-extensions==3.2.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.21.3
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
odata-request-parser==0.1
    # via -r requirements.in
packaging==21.3
    # via drf-yasg
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   django
    #   djangorestframework
    #   drf-yasg
requests==2.28.1
    # via
    #   -r requirements.in
    #   coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sentry-sdk==1.9.5
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.11
    # via
    #   requests
    #   sentry-sdk
